Рефакторинг для требований
