+++ src/components/app/app.tsx
@@ -1,0 +1,30 @@
+import {BrowserRouter, Route, Routes} from 'react-router-dom';
+import {MainScreen} from '../../pages/main-screen/main-screen.tsx';
+import {OfferScreen} from '../../pages/offer-screen/offer-screen.tsx';
+import {NotFoundScreen} from '../../pages/not-found-screen/not-found-screen.tsx';
+import {AppRoute, AuthorizationStatus} from '../../const.ts';
+import {PrivateRoute} from '../private-route/private-route.tsx';
+import {LoginScreen} from '../../pages/login-screen/login-screen.tsx';
+import {FavouritesScreen} from '../../pages/favourites-screen/favourites-screen.tsx';
+
+export function App(): JSX.Element {
+  const authStatus = AuthorizationStatus.Unknown;
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path={AppRoute.Main} element={<MainScreen placesCount={312}/>}/>
+        <Route path={AppRoute.Login} element={<LoginScreen/>}/>
+        <Route path={AppRoute.Offer} element={<OfferScreen/>}/>
+        <Route
+          path={AppRoute.Favorites}
+          element={
+            <PrivateRoute authorizationStatus={authStatus}>
+              <FavouritesScreen/>
+            </PrivateRoute>
+          }
+        />
+        <Route path="*" element={<NotFoundScreen/>}/>
+      </Routes>
+    </BrowserRouter>
+  );
+}
