+++ src/components/card/favourites-card/favourites-card.tsx
@@ -1,42 +1,42 @@
-import {PlaceType} from '../const.ts';
+import {PlaceType} from '../../../const.ts';
 
-type NearCardProps = {
+type FavoritesCardProps = {
   premium?: boolean;
   image: string;
   price: number;
   placeName: string;
   placeType: PlaceType;
-  inBookmarks?: boolean;
 }
 
-export function NearCard({premium, image, price, placeName, placeType, inBookmarks}: NearCardProps): JSX.Element {
+export function FavouritesCard({premium, image, price, placeName, placeType}: FavoritesCardProps): JSX.Element {
   return (
-    <article className="near-places__card place-card">
-      {premium &&
+    <article className="favorites__card place-card">
+      {premium && (
         <div className="place-card__mark">
           <span>Premium</span>
-        </div>}
-      <div className="near-places__image-wrapper place-card__image-wrapper">
+        </div>)}
+      <div className="favorites__image-wrapper place-card__image-wrapper">
         <a href="#">
-          <img className="place-card__image" src={image} width="260" height="200" alt="Place image"/>
+          <img
+            className="place-card__image" src={image} width="150" height="110"
+            alt="Place image"
+          />
         </a>
       </div>
-      <div className="place-card__info">
+      <div className="favorites__card-info place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
             <b className="place-card__price-value">&euro;{price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
           <button
-            className={inBookmarks
-              ? 'place-card__bookmark-button place-card__bookmark-button--active button'
-              : 'place-card__bookmark-button button'}
+            className="place-card__bookmark-button place-card__bookmark-button--active button"
             type="button"
           >
             <svg className="place-card__bookmark-icon" width="18" height="19">
               <use xlinkHref="#icon-bookmark"></use>
             </svg>
-            <span className="visually-hidden">{inBookmarks ? 'In bookmarks' : 'To bookmarks'}</span>
+            <span className="visually-hidden">In bookmarks</span>
           </button>
         </div>
         <div className="place-card__rating rating">
