+++ src/pages/login-screen/login-screen.tsx
@@ -1,11 +1,11 @@
-export function LoginPage(): JSX.Element {
+export function LoginScreen(): JSX.Element {
   return (
     <div className="page page--gray page--login">
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="/">
+              <a className="header__logo-link" href="/public">
                 <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
               </a>
             </div>
