+++ src/pages/not-found-screen/not-found-screen.tsx
@@ -1,9 +1,9 @@
-export function NotFoundPage(): JSX.Element {
+export function NotFoundScreen(): JSX.Element {
   return (
     <center>
       <h1>404</h1>
       <h2>Page not found.</h2>
-      <a href="/">Home page</a>
+      <a href="/public">Home page</a>
     </center>
   );
 }
