+++ src/pages/offer-screen/offer-screen.tsx
@@ -1,14 +1,14 @@
-import {NearCard} from '../components/NearCard.tsx';
-import {PlaceType} from '../const.ts';
+import {NearPlaceCard} from '../../components/card/near-place-card/near-place-card.tsx';
+import {PlaceType} from '../../const.ts';
 
-export function OfferNotLoggedPage(): JSX.Element {
+export function OfferScreen(): JSX.Element {
   return (
     <div className="page">
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="/">
+              <a className="header__logo-link" href="/public">
                 <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
               </a>
             </div>
@@ -18,7 +18,13 @@ export function OfferNotLoggedPage(): JSX.Element {
                   <a className="header__nav-link header__nav-link--profile" href="#">
                     <div className="header__avatar-wrapper user__avatar-wrapper">
                     </div>
-                    <span className="header__login">Sign in</span>
+                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                    <span className="header__favorite-count">3</span>
+                  </a>
+                </li>
+                <li className="header__nav-item">
+                  <a className="header__nav-link" href="#">
+                    <span className="header__signout">Sign out</span>
                   </a>
                 </li>
               </ul>
@@ -183,6 +189,72 @@ export function OfferNotLoggedPage(): JSX.Element {
                     </div>
                   </li>
                 </ul>
+                <form className="reviews__form form" action="#" method="post">
+                  <label className="reviews__label form__label" htmlFor="review">Your review</label>
+                  <div className="reviews__rating-form form__rating">
+                    <input
+                      className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars"
+                      type="radio"
+                    />
+                    <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars"
+                      type="radio"
+                    />
+                    <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars"
+                      type="radio"
+                    />
+                    <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars"
+                      type="radio"
+                    />
+                    <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+
+                    <input
+                      className="form__rating-input visually-hidden" name="rating" value="1" id="1-star"
+                      type="radio"
+                    />
+                    <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
+                      <svg className="form__star-image" width="37" height="33">
+                        <use xlinkHref="#icon-star"></use>
+                      </svg>
+                    </label>
+                  </div>
+                  <textarea
+                    className="reviews__textarea form__textarea" id="review" name="review"
+                    placeholder="Tell how was your stay, what you like and what can be improved"
+                  />
+                  <div className="reviews__button-wrapper">
+                    <p className="reviews__help">
+                      To submit review please make sure to set <span className="reviews__star">rating</span> and
+                      describe
+                      your stay with at least <b className="reviews__text-amount">50 characters</b>.
+                    </p>
+                    <button className="reviews__submit form__submit button" type="submit" disabled>Submit</button>
+                  </div>
+                </form>
               </section>
             </div>
           </div>
@@ -192,20 +264,20 @@ export function OfferNotLoggedPage(): JSX.Element {
           <section className="near-places places">
             <h2 className="near-places__title">Other places in the neighbourhood</h2>
             <div className="near-places__list places__list">
-              <NearCard
+              <NearPlaceCard
                 image={'img/room.jpg'}
                 price={80}
                 placeName={'Wood and stone place'}
                 placeType={PlaceType.Room}
                 inBookmarks
               />
-              <NearCard
+              <NearPlaceCard
                 image={'img/apartment-02.jpg'}
                 price={132}
                 placeName={'Canal View Prinsengracht'}
                 placeType={PlaceType.Apartment}
               />
-              <NearCard
+              <NearPlaceCard
                 premium
                 image={'img/apartment-03.jpg'}
                 price={180}
